task1/pyproject.toml
[tool.poetry]
name = "task1"
version = "0.1.0"
description = ""
authors = ["Liam Vilhelmsson <21284111+Liamwilhelmsson@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.9.2"
python-dotenv = "^1.0.1"
aiohttp = "^3.10.7"
motor = "^3.6.0"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
